Add files via upload
Problem statement
United Nations launched its sustainable development agenda in 2015, reflecting the growing understanding by Member States that a development model that is sustainable for this and future generations offers the best path forward for reducing poverty and improving the lives of people everywhere.Gender disparity a variable measured by gender development index is still a problem globally with the claim of the index being higher presummed average of 0.943. Team atreides has been contracted to figure out how we can predict future GDI index. As the Analytics team we will help provide an indication of the global status in terms of gender development while providing insights into some of the key dimensional index that might drive gender development outcomes.We will be performing hypothesis testing to try to prove or disapprove the claim.